fix(tests): Updated test config for aws secrets and renamed aws tests to reflect purpose.
