+++ api/tests/unit/api/v1/config.py
@@ -181,22 +181,6 @@ password_update_payload = {
     "new_password": "newpassword123",
 }
 
-# Test data for AWS secrets
-aws_secrets_payload: dict[str, Any] = {
-    "provider": "aws",
-    "credentials": {
-        "aws_access_key": "AKIAIOSFODNN7EXAMPLE",
-        "aws_secret_key": "wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY",
-    },
-}
-
-# Test data for Azure secrets
-azure_secrets_payload = {
-    "azure_client_id": "00000000-0000-0000-0000-000000000000",
-    "azure_client_secret": "example-client-secret-value",
-    "azure_tenant_id": "00000000-0000-0000-0000-000000000000",
-    "azure_subscription_id": "00000000-0000-0000-0000-000000000000",
-}
 # ==============================
 #         Job Payloads
 # ==============================
