+++ api/tests/unit/cloud/test_creds_factory.py
@@ -5,7 +5,7 @@ import pytest
 from src.app.cloud.aws_creds import AWSCreds
 from src.app.cloud.creds_factory import CredsFactory
 from src.app.schemas.creds_verify_schema import CredsVerifySchema
-from tests.unit.api.v1.config import aws_secrets_payload
+from tests.common.api.v1.config import aws_secrets_payload
 
 
 def test_creds_factory_non_existent_provider_type() -> None:
